add more test cases
Test a large number of breaks, and interaction with paragraphs
